feat(#84): Add command editor with interactive prompt and mock output
